requirements.txt
pytorch-lightning==1.5.6
numpy==1.25.0
PyYAML==6.0
# torch==1.12.1
# torchvision==0.13.1
matplotlib==3.7.2
h5py==3.9.0
scikit-learn==1.3.0
Pillow==9.4.0
torchinfo==1.8.0
